requirements.txt
amqp==5.0.6
argh==0.26.2
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
dramatiq==1.11.0
ecdsa==0.17.0
fastapi==0.66.0
gevent==21.1.2
greenlet==1.1.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
kombu==5.1.0
lxml==4.6.3
packaging==21.0
passlib==1.7.4
pathtools==0.1.2
pika==1.2.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
watchdog==0.8.3
watchdog-gevent==0.1.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
